STDG-79 fix
 - objects in test-building-data.ts are dedicated
 - remove test-building-data.ts from frontend
 - move test-building-data.ts to snippets folder
 - update dummy data of frontend test webservice
 - restore NavigationPage::showFloor and NavigationPage::showRoute because event triggering from view
