+++ front/src/app/navigation/navigation.page.html
@@ -25,7 +25,7 @@
           <span id="ajaxBtnText">test ajax</span>
         </ion-button>
         <ion-button size="small" color="danger" (click)="fetchRouteToDisplay()">
-          <ion-icon slot="end" azu></ion-icon>
+          <ion-icon slot="end" name="cellular"></ion-icon>
           <span id="ajaxBtnText">test route</span>
         </ion-button>
         <!-- TEST AJAX END-->
