--- front/src/app/navigation/test-building-data.ts
@@ -1,1531 +1,0 @@
-import { Room, PathNode } from '../building-objects-if';
-
-export let testDataRooms: Room[] = [
-  {
-    "name": "KA.304",
-    "sections": [
-      {
-        "Start": { "X": 0, "Y": 0, "Z": 3 },
-        "End": { "X": 150, "Y": 0, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 0, "Z": 3 },
-        "End": { "X": 150, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 100, "Z": 3 },
-        "End": { "X": 0, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 0, "Y": 100, "Z": 3 },
-        "End": { "X": 0, "Y": 0, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 75, "Y": 50, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 135, "Y": 100, "Z": 3 },
-        "End": { "X": 120, "Y": 100, "Z": 3 },
-        "pathNode": { "X": 128, "Y": 100, "Z": 3 }
-      }
-    ],
-    "Color": "#5682a3"
-  },
-  {
-    "name": "KA.302",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 0, "Z": 3 },
-        "End": { "X": 300, "Y": 0, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 0, "Z": 3 },
-        "End": { "X": 300, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 100, "Z": 3 },
-        "End": { "X": 150, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 100, "Z": 3 },
-        "End": { "X": 150, "Y": 0, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 225, "Y": 50, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 180, "Y": 100, "Z": 3 },
-        "End": { "X": 165, "Y": 100, "Z": 3 },
-        "pathNode": { "X": 173, "Y": 100, "Z": 3 }
-      }
-    ],
-    "Color": "#5682a3"
-  },
-  {
-    "name": "Treppe",
-    "sections": [
-      {
-        "Start": { "X": 360, "Y": 0, "Z": 3 },
-        "End": { "X": 400, "Y": 0, "Z": 3 }
-      },
-      {
-        "Start": { "X": 400, "Y": 0, "Z": 3 },
-        "End": { "X": 400, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 400, "Y": 100, "Z": 3 },
-        "End": { "X": 360, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 360, "Y": 100, "Z": 3 },
-        "End": { "X": 360, "Y": 0, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [],
-    "doors": [],
-    "Color": "#AFAFAF"
-  },
-  {
-    "name": "KA.306",
-    "sections": [
-      {
-        "Start": { "X": 0, "Y": 100, "Z": 3 },
-        "End": { "X": 100, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 100, "Y": 100, "Z": 3 },
-        "End": { "X": 100, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 100, "Y": 130, "Z": 3 },
-        "End": { "X": 0, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 0, "Y": 130, "Z": 3 },
-        "End": { "X": 0, "Y": 100, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 50, "Y": 115, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 100, "Y": 110, "Z": 3 },
-        "End": { "X": 100, "Y": 125, "Z": 3 },
-        "pathNode": { "X": 100, "Y": 118, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "corridor",
-    "sections": [
-      {
-        "Start": { "X": 100, "Y": 100, "Z": 3 },
-        "End": { "X": 300, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 100, "Z": 3 },
-        "End": { "X": 300, "Y": 0, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 0, "Z": 3 },
-        "End": { "X": 360, "Y": 0, "Z": 3 }
-      },
-      {
-        "Start": { "X": 360, "Y": 0, "Z": 3 },
-        "End": { "X": 360, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 360, "Y": 130, "Z": 3 },
-        "End": { "X": 290, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 130, "Z": 3 },
-        "End": { "X": 290, "Y": 155, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 155, "Z": 3 },
-        "End": { "X": 225, "Y": 155, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 155, "Z": 3 },
-        "End": { "X": 225, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 130, "Z": 3 },
-        "End": { "X": 100, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 100, "Y": 130, "Z": 3 },
-        "End": { "X": 100, "Y": 100, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 128, "Y": 115, "Z": 3 },
-      { "X": 173, "Y": 115, "Z": 3 },
-      { "X": 257, "Y": 115, "Z": 3 }
-    ],
-    "doors": [],
-    "Color": "#BBB"
-  },
-  {
-    "name": "Aufzug",
-    "sections": [
-      {
-        "Start": { "X": 360, "Y": 100, "Z": 3 },
-        "End": { "X": 400, "Y": 100, "Z": 3 }
-      },
-      {
-        "Start": { "X": 400, "Y": 100, "Z": 3 },
-        "End": { "X": 400, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 400, "Y": 130, "Z": 3 },
-        "End": { "X": 360, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 360, "Y": 130, "Z": 3 },
-        "End": { "X": 360, "Y": 100, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [],
-    "doors": [],
-    "Color": "#817464"
-  },
-  {
-    "name": "KA.308",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 130, "Z": 3 },
-        "End": { "X": 225, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 130, "Z": 3 },
-        "End": { "X": 225, "Y": 210, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 210, "Z": 3 },
-        "End": { "X": 150, "Y": 210, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 210, "Z": 3 },
-        "End": { "X": 150, "Y": 130, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 187, "Y": 170, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 225, "Y": 190, "Z": 3 },
-        "End": { "X": 225, "Y": 205, "Z": 3 },
-        "pathNode": { "X": 225, "Y": 198, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "Treppe",
-    "sections": [
-      {
-        "Start": { "X": 225, "Y": 155, "Z": 3 },
-        "End": { "X": 290, "Y": 155, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 155, "Z": 3 },
-        "End": { "X": 290, "Y": 185, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 185, "Z": 3 },
-        "End": { "X": 225, "Y": 185, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 185, "Z": 3 },
-        "End": { "X": 225, "Y": 155, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 257, "Y": 170, "Z": 3 }
-    ],
-    "doors": [],
-    "Color": "#AFAFAF"
-  },
-  {
-    "name": "WC Herren",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 130, "Z": 3 },
-        "End": { "X": 380, "Y": 130, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 130, "Z": 3 },
-        "End": { "X": 380, "Y": 175, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 175, "Z": 3 },
-        "End": { "X": 320, "Y": 175, "Z": 3 }
-      },
-      {
-        "Start": { "X": 320, "Y": 175, "Z": 3 },
-        "End": { "X": 310, "Y": 195, "Z": 3 }
-      },
-      {
-        "Start": { "X": 310, "Y": 195, "Z": 3 },
-        "End": { "X": 290, "Y": 195, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 195, "Z": 3 },
-        "End": { "X": 290, "Y": 130, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 160, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 325, "Y": 175, "Z": 3 },
-        "End": { "X": 340, "Y": 175, "Z": 3 },
-        "pathNode": { "X": 333, "Y": 175, "Z": 3 }
-      }
-    ],
-    "Color": "#AEA693"
-  },
-  {
-    "name": "WC Damen",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 200, "Z": 3 },
-        "End": { "X": 310, "Y": 200, "Z": 3 }
-      },
-      {
-        "Start": { "X": 310, "Y": 200, "Z": 3 },
-        "End": { "X": 320, "Y": 180, "Z": 3 }
-      },
-      {
-        "Start": { "X": 320, "Y": 180, "Z": 3 },
-        "End": { "X": 380, "Y": 180, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 180, "Z": 3 },
-        "End": { "X": 380, "Y": 220, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 220, "Z": 3 },
-        "End": { "X": 290, "Y": 220, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 220, "Z": 3 },
-        "End": { "X": 290, "Y": 200, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 200, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 325, "Y": 180, "Z": 3 },
-        "End": { "X": 340, "Y": 180, "Z": 3 },
-        "pathNode": { "X": 333, "Y": 180, "Z": 3 }
-      }
-    ],
-    "Color": "#AEA693"
-  },
-  {
-    "name": "KA.310",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 210, "Z": 3 },
-        "End": { "X": 225, "Y": 210, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 210, "Z": 3 },
-        "End": { "X": 225, "Y": 235, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 235, "Z": 3 },
-        "End": { "X": 150, "Y": 235, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 235, "Z": 3 },
-        "End": { "X": 150, "Y": 210, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 187, "Y": 222, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 225, "Y": 215, "Z": 3 },
-        "End": { "X": 225, "Y": 225, "Z": 3 },
-        "pathNode": { "X": 225, "Y": 220, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.312",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 235, "Z": 3 },
-        "End": { "X": 225, "Y": 235, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 235, "Z": 3 },
-        "End": { "X": 225, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 265, "Z": 3 },
-        "End": { "X": 150, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 265, "Z": 3 },
-        "End": { "X": 150, "Y": 235, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 187, "Y": 250, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 225, "Y": 240, "Z": 3 },
-        "End": { "X": 225, "Y": 250, "Z": 3 },
-        "pathNode": { "X": 225, "Y": 245, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.305",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 245, "Z": 3 },
-        "End": { "X": 380, "Y": 235, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 235, "Z": 3 },
-        "End": { "X": 380, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 265, "Z": 3 },
-        "End": { "X": 290, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 265, "Z": 3 },
-        "End": { "X": 290, "Y": 245, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 250, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 325, "Y": 241, "Z": 3 },
-        "End": { "X": 345, "Y": 239, "Z": 3 },
-        "pathNode": { "X": 335, "Y": 240, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.314",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 265, "Z": 3 },
-        "End": { "X": 240, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 265, "Z": 3 },
-        "End": { "X": 240, "Y": 295, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 295, "Z": 3 },
-        "End": { "X": 150, "Y": 295, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 295, "Z": 3 },
-        "End": { "X": 150, "Y": 265, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 280, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 275, "Z": 3 },
-        "End": { "X": 240, "Y": 285, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 280, "Z": 3 }
-      },
-      {
-        "Start": { "X": 205, "Y": 295, "Z": 3 },
-        "End": { "X": 220, "Y": 295, "Z": 3 },
-        "pathNode": { "X": 213, "Y": 295, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.316",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 295, "Z": 3 },
-        "End": { "X": 240, "Y": 295, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 295, "Z": 3 },
-        "End": { "X": 240, "Y": 325, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 325, "Z": 3 },
-        "End": { "X": 150, "Y": 325, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 325, "Z": 3 },
-        "End": { "X": 150, "Y": 295, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 310, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 305, "Z": 3 },
-        "End": { "X": 240, "Y": 315, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 310, "Z": 3 }
-      },
-      {
-        "Start": { "X": 205, "Y": 295, "Z": 3 },
-        "End": { "X": 220, "Y": 295, "Z": 3 },
-        "pathNode": { "X": 213, "Y": 295, "Z": 3 }
-      },
-      {
-        "Start": { "X": 210, "Y": 325, "Z": 3 },
-        "End": { "X": 225, "Y": 325, "Z": 3 },
-        "pathNode": { "X": 218, "Y": 325, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.318",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 325, "Z": 3 },
-        "End": { "X": 240, "Y": 325, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 325, "Z": 3 },
-        "End": { "X": 240, "Y": 350, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 350, "Z": 3 },
-        "End": { "X": 150, "Y": 350, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 350, "Z": 3 },
-        "End": { "X": 150, "Y": 325, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 337, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 330, "Z": 3 },
-        "End": { "X": 240, "Y": 340, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 335, "Z": 3 }
-      },
-      {
-        "Start": { "X": 210, "Y": 325, "Z": 3 },
-        "End": { "X": 225, "Y": 325, "Z": 3 },
-        "pathNode": { "X": 218, "Y": 325, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.320",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 350, "Z": 3 },
-        "End": { "X": 240, "Y": 350, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 350, "Z": 3 },
-        "End": { "X": 240, "Y": 380, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 380, "Z": 3 },
-        "End": { "X": 150, "Y": 380, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 380, "Z": 3 },
-        "End": { "X": 150, "Y": 350, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 365, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 360, "Z": 3 },
-        "End": { "X": 240, "Y": 370, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 365, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.322",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 380, "Z": 3 },
-        "End": { "X": 240, "Y": 380, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 380, "Z": 3 },
-        "End": { "X": 240, "Y": 460, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 460, "Z": 3 },
-        "End": { "X": 150, "Y": 460, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 460, "Z": 3 },
-        "End": { "X": 150, "Y": 380, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 420, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 440, "Z": 3 },
-        "End": { "X": 240, "Y": 450, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 445, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.324",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 460, "Z": 3 },
-        "End": { "X": 240, "Y": 460, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 460, "Z": 3 },
-        "End": { "X": 240, "Y": 490, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 490, "Z": 3 },
-        "End": { "X": 150, "Y": 490, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 490, "Z": 3 },
-        "End": { "X": 150, "Y": 460, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 475, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 465, "Z": 3 },
-        "End": { "X": 240, "Y": 475, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 470, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.326",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 490, "Z": 3 },
-        "End": { "X": 240, "Y": 490, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 490, "Z": 3 },
-        "End": { "X": 240, "Y": 520, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 520, "Z": 3 },
-        "End": { "X": 150, "Y": 520, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 520, "Z": 3 },
-        "End": { "X": 150, "Y": 490, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 505, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 500, "Z": 3 },
-        "End": { "X": 240, "Y": 510, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 505, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.328",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 520, "Z": 3 },
-        "End": { "X": 240, "Y": 520, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 520, "Z": 3 },
-        "End": { "X": 240, "Y": 550, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 550, "Z": 3 },
-        "End": { "X": 150, "Y": 550, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 550, "Z": 3 },
-        "End": { "X": 150, "Y": 520, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 535, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 530, "Z": 3 },
-        "End": { "X": 240, "Y": 540, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 535, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.307",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 265, "Z": 3 },
-        "End": { "X": 380, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 265, "Z": 3 },
-        "End": { "X": 380, "Y": 320, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 320, "Z": 3 },
-        "End": { "X": 290, "Y": 320, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 320, "Z": 3 },
-        "End": { "X": 290, "Y": 265, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 292, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 280, "Z": 3 },
-        "End": { "X": 290, "Y": 295, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 288, "Z": 3 }
-      },
-      {
-        "Start": { "X": 305, "Y": 320, "Z": 3 },
-        "End": { "X": 315, "Y": 320, "Z": 3 },
-        "pathNode": { "X": 310, "Y": 320, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.309",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 320, "Z": 3 },
-        "End": { "X": 380, "Y": 320, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 320, "Z": 3 },
-        "End": { "X": 380, "Y": 420, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 420, "Z": 3 },
-        "End": { "X": 290, "Y": 420, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 420, "Z": 3 },
-        "End": { "X": 290, "Y": 320, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 370, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 330, "Z": 3 },
-        "End": { "X": 290, "Y": 345, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 338, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 395, "Z": 3 },
-        "End": { "X": 290, "Y": 410, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 403, "Z": 3 }
-      },
-      {
-        "Start": { "X": 305, "Y": 320, "Z": 3 },
-        "End": { "X": 315, "Y": 320, "Z": 3 },
-        "pathNode": { "X": 310, "Y": 320, "Z": 3 }
-      },
-      {
-        "Start": { "X": 305, "Y": 420, "Z": 3 },
-        "End": { "X": 315, "Y": 420, "Z": 3 },
-        "pathNode": { "X": 310, "Y": 420, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.311",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 420, "Z": 3 },
-        "End": { "X": 380, "Y": 420, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 420, "Z": 3 },
-        "End": { "X": 380, "Y": 520, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 520, "Z": 3 },
-        "End": { "X": 290, "Y": 520, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 520, "Z": 3 },
-        "End": { "X": 290, "Y": 420, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 470, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 430, "Z": 3 },
-        "End": { "X": 290, "Y": 445, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 438, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 495, "Z": 3 },
-        "End": { "X": 290, "Y": 510, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 503, "Z": 3 }
-      },
-      {
-        "Start": { "X": 305, "Y": 420, "Z": 3 },
-        "End": { "X": 315, "Y": 420, "Z": 3 },
-        "pathNode": { "X": 310, "Y": 420, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 520, "Z": 3 },
-        "End": { "X": 380, "Y": 520, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 520, "Z": 3 },
-        "End": { "X": 380, "Y": 550, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 550, "Z": 3 },
-        "End": { "X": 290, "Y": 550, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 550, "Z": 3 },
-        "End": { "X": 290, "Y": 520, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 535, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 530, "Z": 3 },
-        "End": { "X": 290, "Y": 540, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 535, "Z": 3 }
-      }
-    ],
-    "Color": "#949494"
-  },
-  {
-    "name": "Treppe",
-    "sections": [
-      {
-        "Start": { "X": 130, "Y": 550, "Z": 3 },
-        "End": { "X": 240, "Y": 550, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 550, "Z": 3 },
-        "End": { "X": 240, "Y": 590, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 590, "Z": 3 },
-        "End": { "X": 130, "Y": 590, "Z": 3 }
-      },
-      {
-        "Start": { "X": 130, "Y": 590, "Z": 3 },
-        "End": { "X": 130, "Y": 550, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [],
-    "doors": [],
-    "Color": "#AFAFAF"
-  },
-  {
-    "name": "WC Damen",
-    "sections": [
-      {
-        "Start": { "X": 130, "Y": 590, "Z": 3 },
-        "End": { "X": 190, "Y": 590, "Z": 3 }
-      },
-      {
-        "Start": { "X": 190, "Y": 590, "Z": 3 },
-        "End": { "X": 190, "Y": 610, "Z": 3 }
-      },
-      {
-        "Start": { "X": 190, "Y": 610, "Z": 3 },
-        "End": { "X": 170, "Y": 610, "Z": 3 }
-      },
-      {
-        "Start": { "X": 170, "Y": 610, "Z": 3 },
-        "End": { "X": 170, "Y": 620, "Z": 3 }
-      },
-      {
-        "Start": { "X": 170, "Y": 620, "Z": 3 },
-        "End": { "X": 130, "Y": 620, "Z": 3 }
-      },
-      {
-        "Start": { "X": 130, "Y": 620, "Z": 3 },
-        "End": { "X": 130, "Y": 590, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 160, "Y": 605, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 185, "Y": 610, "Z": 3 },
-        "End": { "X": 175, "Y": 610, "Z": 3 },
-        "pathNode": { "X": 180, "Y": 610, "Z": 3 }
-      }
-    ],
-    "Color": "#AEA693"
-  },
-  {
-    "name": "WC Herren",
-    "sections": [
-      {
-        "Start": { "X": 130, "Y": 620, "Z": 3 },
-        "End": { "X": 170, "Y": 620, "Z": 3 }
-      },
-      {
-        "Start": { "X": 170, "Y": 620, "Z": 3 },
-        "End": { "X": 170, "Y": 630, "Z": 3 }
-      },
-      {
-        "Start": { "X": 170, "Y": 630, "Z": 3 },
-        "End": { "X": 190, "Y": 630, "Z": 3 }
-      },
-      {
-        "Start": { "X": 190, "Y": 630, "Z": 3 },
-        "End": { "X": 190, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 190, "Y": 650, "Z": 3 },
-        "End": { "X": 130, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 130, "Y": 650, "Z": 3 },
-        "End": { "X": 130, "Y": 620, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 160, "Y": 635, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 185, "Y": 630, "Z": 3 },
-        "End": { "X": 175, "Y": 630, "Z": 3 },
-        "pathNode": { "X": 180, "Y": 630, "Z": 3 }
-      }
-    ],
-    "Color": "#AEA693"
-  },
-  {
-    "name": "Aufzug",
-    "sections": [
-      {
-        "Start": { "X": 190, "Y": 590, "Z": 3 },
-        "End": { "X": 240, "Y": 590, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 590, "Z": 3 },
-        "End": { "X": 240, "Y": 610, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 610, "Z": 3 },
-        "End": { "X": 190, "Y": 610, "Z": 3 }
-      },
-      {
-        "Start": { "X": 190, "Y": 610, "Z": 3 },
-        "End": { "X": 190, "Y": 590, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [],
-    "doors": [],
-    "Color": "#817464"
-  },
-  {
-    "name": "corridor",
-    "sections": [
-      {
-        "Start": { "X": 225, "Y": 185, "Z": 3 },
-        "End": { "X": 290, "Y": 185, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 185, "Z": 3 },
-        "End": { "X": 290, "Y": 195, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 195, "Z": 3 },
-        "End": { "X": 310, "Y": 195, "Z": 3 }
-      },
-      {
-        "Start": { "X": 310, "Y": 195, "Z": 3 },
-        "End": { "X": 320, "Y": 175, "Z": 3 }
-      },
-      {
-        "Start": { "X": 320, "Y": 175, "Z": 3 },
-        "End": { "X": 380, "Y": 175, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 175, "Z": 3 },
-        "End": { "X": 380, "Y": 180, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 180, "Z": 3 },
-        "End": { "X": 320, "Y": 180, "Z": 3 }
-      },
-      {
-        "Start": { "X": 320, "Y": 180, "Z": 3 },
-        "End": { "X": 310, "Y": 200, "Z": 3 }
-      },
-      {
-        "Start": { "X": 310, "Y": 200, "Z": 3 },
-        "End": { "X": 290, "Y": 200, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 200, "Z": 3 },
-        "End": { "X": 290, "Y": 220, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 220, "Z": 3 },
-        "End": { "X": 380, "Y": 220, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 220, "Z": 3 },
-        "End": { "X": 380, "Y": 235, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 235, "Z": 3 },
-        "End": { "X": 290, "Y": 245, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 245, "Z": 3 },
-        "End": { "X": 290, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 930, "Z": 3 },
-        "End": { "X": 310, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 310, "Y": 930, "Z": 3 },
-        "End": { "X": 310, "Y": 1080, "Z": 3 }
-      },
-      {
-        "Start": { "X": 310, "Y": 1080, "Z": 3 },
-        "End": { "X": 220, "Y": 1080, "Z": 3 }
-      },
-      {
-        "Start": { "X": 220, "Y": 1080, "Z": 3 },
-        "End": { "X": 220, "Y": 1050, "Z": 3 }
-      },
-      {
-        "Start": { "X": 220, "Y": 1050, "Z": 3 },
-        "End": { "X": 245, "Y": 1050, "Z": 3 }
-      },
-      {
-        "Start": { "X": 245, "Y": 1050, "Z": 3 },
-        "End": { "X": 245, "Y": 960, "Z": 3 }
-      },
-      {
-        "Start": { "X": 245, "Y": 960, "Z": 3 },
-        "End": { "X": 220, "Y": 960, "Z": 3 }
-      },
-      {
-        "Start": { "X": 220, "Y": 960, "Z": 3 },
-        "End": { "X": 220, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 220, "Y": 930, "Z": 3 },
-        "End": { "X": 240, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 930, "Z": 3 },
-        "End": { "X": 240, "Y": 630, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 630, "Z": 3 },
-        "End": { "X": 170, "Y": 630, "Z": 3 }
-      },
-      {
-        "Start": { "X": 170, "Y": 630, "Z": 3 },
-        "End": { "X": 170, "Y": 610, "Z": 3 }
-      },
-      {
-        "Start": { "X": 170, "Y": 610, "Z": 3 },
-        "End": { "X": 240, "Y": 610, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 610, "Z": 3 },
-        "End": { "X": 240, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 265, "Z": 3 },
-        "End": { "X": 225, "Y": 265, "Z": 3 }
-      },
-      {
-        "Start": { "X": 225, "Y": 265, "Z": 3 },
-        "End": { "X": 225, "Y": 185, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 257, "Y": 198, "Z": 3 },
-      { "X": 257, "Y": 222, "Z": 3 },
-      { "X": 257, "Y": 245, "Z": 3 },
-      { "X": 265, "Y": 280, "Z": 3 },
-      { "X": 265, "Y": 288, "Z": 3 },
-      { "X": 265, "Y": 310, "Z": 3 },
-      { "X": 265, "Y": 335, "Z": 3 },
-      { "X": 265, "Y": 365, "Z": 3 },
-      { "X": 265, "Y": 403, "Z": 3 },
-      { "X": 265, "Y": 438, "Z": 3 },
-      { "X": 265, "Y": 445, "Z": 3 },
-      { "X": 265, "Y": 470, "Z": 3 },
-      { "X": 265, "Y": 503, "Z": 3 },
-      { "X": 265, "Y": 505, "Z": 3 },
-      { "X": 265, "Y": 535, "Z": 3 },
-      { "X": 180, "Y": 620, "Z": 3 },
-      { "X": 265, "Y": 620, "Z": 3 },
-      { "X": 265, "Y": 628, "Z": 3 },
-      { "X": 265, "Y": 673, "Z": 3 },
-      { "X": 265, "Y": 728, "Z": 3 },
-      { "X": 265, "Y": 773, "Z": 3 },
-      { "X": 265, "Y": 828, "Z": 3 },
-      { "X": 265, "Y": 870, "Z": 3 },
-      { "X": 265, "Y": 903, "Z": 3 },
-      { "X": 265, "Y": 910, "Z": 3 },
-      { "X": 310, "Y": 198, "Z": 3 },
-      { "X": 320, "Y": 178, "Z": 3 },
-      { "X": 333, "Y": 178, "Z": 3 },
-      { "X": 335, "Y": 230, "Z": 3 },
-      { "X": 257, "Y": 230, "Z": 3 }
-    ],
-    "doors": [],
-    "Color": "#BBB"
-  },
-  {
-    "name": "",
-    "sections": [
-      {
-        "Start": { "X": 190, "Y": 630, "Z": 3 },
-        "End": { "X": 240, "Y": 630, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 630, "Z": 3 },
-        "End": { "X": 240, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 650, "Z": 3 },
-        "End": { "X": 190, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 190, "Y": 650, "Z": 3 },
-        "End": { "X": 190, "Y": 630, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [],
-    "doors": [],
-    "Color": "#949494"
-  },
-  {
-    "name": "KA.334",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 650, "Z": 3 },
-        "End": { "X": 240, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 650, "Z": 3 },
-        "End": { "X": 240, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 750, "Z": 3 },
-        "End": { "X": 150, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 750, "Z": 3 },
-        "End": { "X": 150, "Y": 650, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-        { "X": 195, "Y": 700, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 665, "Z": 3 },
-        "End": { "X": 240, "Y": 680, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 673, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 720, "Z": 3 },
-        "End": { "X": 240, "Y": 735, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 728, "Z": 3 }
-      }
-    ],
-    "Color": "#5682a3"
-  },
-  {
-    "name": "KA.336",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 750, "Z": 3 },
-        "End": { "X": 240, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 750, "Z": 3 },
-        "End": { "X": 240, "Y": 830, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 830, "Z": 3 },
-        "End": { "X": 150, "Y": 830, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 830, "Z": 3 },
-        "End": { "X": 150, "Y": 750, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 790, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 765, "Z": 3 },
-        "End": { "X": 240, "Y": 780, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 773, "Z": 3 }
-      }
-    ],
-    "Color": "#5682a3"
-  },
-  {
-    "name": "KA.338",
-    "sections": [
-      {
-        "Start": { "X": 150, "Y": 830, "Z": 3 },
-        "End": { "X": 240, "Y": 830, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 830, "Z": 3 },
-        "End": { "X": 240, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 930, "Z": 3 },
-        "End": { "X": 150, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 150, "Y": 930, "Z": 3 },
-        "End": { "X": 150, "Y": 830, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 195, "Y": 880, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 240, "Y": 865, "Z": 3 },
-        "End": { "X": 240, "Y": 875, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 870, "Z": 3 }
-      },
-      {
-        "Start": { "X": 240, "Y": 905, "Z": 3 },
-        "End": { "X": 240, "Y": 915, "Z": 3 },
-        "pathNode": { "X": 240, "Y": 910, "Z": 3 }
-      }
-    ],
-    "Color": "#5682a3"
-  },
-  {
-    "name": "KA.313",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 550, "Z": 3 },
-        "End": { "X": 400, "Y": 550, "Z": 3 }
-      },
-      {
-        "Start": { "X": 400, "Y": 550, "Z": 3 },
-        "End": { "X": 400, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 400, "Y": 650, "Z": 3 },
-        "End": { "X": 290, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 650, "Z": 3 },
-        "End": { "X": 290, "Y": 550, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 345, "Y": 600, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 620, "Z": 3 },
-        "End": { "X": 290, "Y": 635, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 628, "Z": 3 }
-      }
-    ],
-    "Color": "#5682a3"
-  },
-  {
-    "name": "KA.315",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 650, "Z": 3 },
-        "End": { "X": 380, "Y": 650, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 650, "Z": 3 },
-        "End": { "X": 380, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 750, "Z": 3 },
-        "End": { "X": 290, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 750, "Z": 3 },
-        "End": { "X": 290, "Y": 650, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 700, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 665, "Z": 3 },
-        "End": { "X": 290, "Y": 680, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 673, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 720, "Z": 3 },
-        "End": { "X": 290, "Y": 735, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 728, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 750, "Z": 3 },
-        "End": { "X": 315, "Y": 750, "Z": 3 },
-        "pathNode": { "X": 308, "Y": 750, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.317",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 750, "Z": 3 },
-        "End": { "X": 380, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 750, "Z": 3 },
-        "End": { "X": 380, "Y": 850, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 850, "Z": 3 },
-        "End": { "X": 290, "Y": 850, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 850, "Z": 3 },
-        "End": { "X": 290, "Y": 750, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 800, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 765, "Z": 3 },
-        "End": { "X": 290, "Y": 780, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 773, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 820, "Z": 3 },
-        "End": { "X": 290, "Y": 835, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 828, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 750, "Z": 3 },
-        "End": { "X": 315, "Y": 750, "Z": 3 },
-        "pathNode": { "X": 308, "Y": 750, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 850, "Z": 3 },
-        "End": { "X": 315, "Y": 850, "Z": 3 },
-        "pathNode": { "X": 308, "Y": 850, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "KA.319",
-    "sections": [
-      {
-        "Start": { "X": 290, "Y": 850, "Z": 3 },
-        "End": { "X": 380, "Y": 850, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 850, "Z": 3 },
-        "End": { "X": 380, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 380, "Y": 930, "Z": 3 },
-        "End": { "X": 290, "Y": 930, "Z": 3 }
-      },
-      {
-        "Start": { "X": 290, "Y": 930, "Z": 3 },
-        "End": { "X": 290, "Y": 850, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [
-      { "X": 335, "Y": 890, "Z": 3 }
-    ],
-    "doors": [
-      {
-        "Start": { "X": 290, "Y": 895, "Z": 3 },
-        "End": { "X": 290, "Y": 910, "Z": 3 },
-        "pathNode": { "X": 290, "Y": 903, "Z": 3 }
-      },
-      {
-        "Start": { "X": 300, "Y": 850, "Z": 3 },
-        "End": { "X": 315, "Y": 850, "Z": 3 },
-        "pathNode": { "X": 308, "Y": 850, "Z": 3 }
-      }
-    ],
-    "Color": "#4b7989"
-  },
-  {
-    "name": "Treppe",
-    "sections": [
-      {
-        "Start": { "X": 220, "Y": 960, "Z": 3 },
-        "End": { "X": 245, "Y": 960, "Z": 3 }
-      },
-      {
-        "Start": { "X": 245, "Y": 960, "Z": 3 },
-        "End": { "X": 245, "Y": 1050, "Z": 3 }
-      },
-      {
-        "Start": { "X": 245, "Y": 1050, "Z": 3 },
-        "End": { "X": 220, "Y": 1050, "Z": 3 }
-      },
-      {
-        "Start": { "X": 220, "Y": 1050, "Z": 3 },
-        "End": { "X": 220, "Y": 960, "Z": 3 }
-      }
-    ],
-    "alias": [],
-    "pathNodes": [],
-    "doors": [],
-    "Color": "#AFAFAF"
-  }
-];
-
-export let testDataPathNodes:PathNode[] = [
-  { "X": 187, "Y": 170, "Z": 3 },
-  { "X": 225, "Y": 198, "Z": 3 },
-  { "X": 257, "Y": 198, "Z": 3 },
-  { "X": 257, "Y": 222, "Z": 3 },
-  { "X": 257, "Y": 230, "Z": 3 },
-  { "X": 257, "Y": 245, "Z": 3 },
-  { "X": 265, "Y": 280, "Z": 3 },
-  { "X": 265, "Y": 288, "Z": 3 },
-  { "X": 265, "Y": 310, "Z": 3 },
-  { "X": 265, "Y": 335, "Z": 3 },
-  { "X": 265, "Y": 365, "Z": 3 },
-  { "X": 265, "Y": 403, "Z": 3 },
-  { "X": 265, "Y": 438, "Z": 3 },
-  { "X": 265, "Y": 445, "Z": 3 },
-  { "X": 265, "Y": 470, "Z": 3 },
-  { "X": 265, "Y": 503, "Z": 3 },
-  { "X": 265, "Y": 505, "Z": 3 },
-  { "X": 265, "Y": 535, "Z": 3 },
-  { "X": 265, "Y": 620, "Z": 3 },
-  { "X": 265, "Y": 628, "Z": 3 },
-  { "X": 290, "Y": 628, "Z": 3 },
-  { "X": 345, "Y": 600, "Z": 3 }
-];
