+++ snippets/httpServer/dummy-route.json
@@ -18,15 +18,6 @@
     "Group": null
   },
   {
-    "Id": 4,
-    "Coordinate": {
-      "X": 257,
-      "Y": 170,
-      "Z": 3
-    },
-    "Group": null
-  },
-  {
     "Id": 14,
     "Coordinate": {
       "X": 257,
